[client] wayland: split xdg-shell handling into a module
This will allow libdecor to be added as a new pseudo-shell.
